+++ IsekaiMod/Content/Features/IsekaiProtagonist/Archetypes/Mastermind/MastermindQuickFooted.cs
@@ -16,7 +16,7 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.Archetypes.Mastermind {
         public static void Add() {
             var MastermindQuickFooted = Helpers.CreateBlueprint<BlueprintFeature>(IsekaiContext, "MastermindQuickFooted", bp => {
                 bp.SetName(IsekaiContext, "Quick-Footed");
-                bp.SetDescription(IsekaiContext, "At 16th level, you gain a competence {g|Encyclopedia:Bonus}bonus{/g} to your {g|Encyclopedia:Initiative}initiative{/g} "
+                bp.SetDescription(IsekaiContext, "At 15th level, you gain a competence {g|Encyclopedia:Bonus}bonus{/g} to your {g|Encyclopedia:Initiative}initiative{/g} "
                     + "{g|Encyclopedia:Check}checks{/g} equal to your {g|Encyclopedia:Intelligence}Intelligence{/g} modifier.");
                 bp.m_Icon = Icon_ExpeditiousRetreat;
                 bp.AddComponent<DerivativeStatBonus>(c => {