+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Buffed `God Emperor` to get another OP ability at 5th and 15th level.
 - Buffed `Hero` to get another Special Power at 3rd and 17th level.
 - Buffed `Hero` to get Tower shield and exotic weapon proficiency.
+- Changed `Isekai Protagonist` `Otherworldly Stamina` obtained from 15 -> 13.
+- Changed `Isekai Protagonist` `Quick-Footed` obtained from 16 -> 15.
 ### Removed
 - Removed `Villain` Archetype (Isekai Protagonist).
 - Removed `Kinetic Power` Special Power.